# backend/src/main/kotlin/com/thredge/backend/api/ThreadController.kt
package com.thredge.backend.api

import com.thredge.backend.api.dto.EntryDetail
import com.thredge.backend.api.dto.EntryRequest
import com.thredge.backend.api.dto.PageResponse
import com.thredge.backend.api.dto.ThreadCreateRequest
import com.thredge.backend.api.dto.ThreadDetail
import com.thredge.backend.api.dto.ThreadFeedItem
import com.thredge.backend.api.dto.ThreadSummary
import com.thredge.backend.api.dto.ThreadUpdateRequest
import com.thredge.backend.service.ThreadService
import com.thredge.backend.support.AuthSupport
import com.thredge.backend.support.PagingDefaults
import com.thredge.backend.support.ValidationMessages
import jakarta.validation.Valid
import jakarta.validation.constraints.Max
import jakarta.validation.constraints.Min
import jakarta.validation.constraints.NotBlank
import java.time.LocalDate
import org.springframework.data.domain.PageRequest
import org.springframework.format.annotation.DateTimeFormat
import org.springframework.security.core.Authentication
import org.springframework.validation.annotation.Validated
import org.springframework.web.bind.annotation.DeleteMapping
import org.springframework.web.bind.annotation.GetMapping
import org.springframework.web.bind.annotation.PatchMapping
import org.springframework.web.bind.annotation.PathVariable
import org.springframework.web.bind.annotation.PostMapping
import org.springframework.web.bind.annotation.RequestBody
import org.springframework.web.bind.annotation.RequestMapping
import org.springframework.web.bind.annotation.RequestParam
import org.springframework.web.bind.annotation.RestController
import org.springframework.http.ResponseEntity

@RestController
@RequestMapping("/api/threads")
@Validated
class ThreadController(
        private val threadService: ThreadService,
        private val authSupport: AuthSupport,
) {
        @GetMapping
        fun list(
                @RequestParam(defaultValue = "${PagingDefaults.DEFAULT_PAGE}") @Min(0L) page: Int,
                @RequestParam(defaultValue = "${PagingDefaults.THREAD_DEFAULT_SIZE}")
                @Min(1L)
                @Max(PagingDefaults.THREAD_MAX_SIZE)
                size: Int,
                authentication: Authentication?,
        ): PageResponse<ThreadSummary> {
                val ownerUsername = authSupport.requireUsername(authentication)
                return threadService.list(ownerUsername, PageRequest.of(page, size))
        }

        @GetMapping("/feed")
        fun feed(
                @RequestParam(defaultValue = "${PagingDefaults.DEFAULT_PAGE}") @Min(0L) page: Int,
                @RequestParam(defaultValue = "${PagingDefaults.THREAD_DEFAULT_SIZE}")
                @Min(1L)
                @Max(PagingDefaults.THREAD_MAX_SIZE)
                size: Int,
                @RequestParam(required = false)
                @DateTimeFormat(iso = DateTimeFormat.ISO.DATE)
                date: LocalDate?,
                @RequestParam(required = false) categoryIds: List<String>?,
                authentication: Authentication?,
        ): PageResponse<ThreadFeedItem> {
                val ownerUsername = authSupport.requireUsername(authentication)
                return threadService.feed(
                        ownerUsername,
                        PageRequest.of(page, size),
                        date,
                        categoryIds
                )
        }

        @GetMapping("/search")
        fun searchThreads(
                @RequestParam @NotBlank(message = ValidationMessages.QUERY_REQUIRED) query: String,
                @RequestParam(required = false) categoryIds: List<String>?,
                @RequestParam(required = false)
                @DateTimeFormat(iso = DateTimeFormat.ISO.DATE)
                date: LocalDate?,
                @RequestParam(defaultValue = "${PagingDefaults.DEFAULT_PAGE}") @Min(0L) page: Int,
                @RequestParam(defaultValue = "${PagingDefaults.THREAD_DEFAULT_SIZE}")
                @Min(1L)
                @Max(PagingDefaults.THREAD_MAX_SIZE)
                size: Int,
                authentication: Authentication?,
        ): PageResponse<ThreadFeedItem> {
                val ownerUsername = authSupport.requireUsername(authentication)
                return threadService.searchThreads(
                        ownerUsername,
                        query,
                        categoryIds,
                        date,
                        PageRequest.of(page, size)
                )
        }

        @GetMapping("/hidden")
        fun listHidden(
                @RequestParam(defaultValue = "${PagingDefaults.DEFAULT_PAGE}") @Min(0L) page: Int,
                @RequestParam(defaultValue = "${PagingDefaults.THREAD_DEFAULT_SIZE}")
                @Min(1L)
                @Max(PagingDefaults.THREAD_MAX_SIZE)
                size: Int,
                authentication: Authentication?,
        ): PageResponse<ThreadSummary> {
                val ownerUsername = authSupport.requireUsername(authentication)
                return threadService.listHidden(ownerUsername, PageRequest.of(page, size))
        }

        @GetMapping("/hidden/search")
        fun searchHiddenThreads(
                @RequestParam @NotBlank(message = ValidationMessages.QUERY_REQUIRED) query: String,
                @RequestParam(required = false) categoryIds: List<String>?,
                @RequestParam(defaultValue = "${PagingDefaults.DEFAULT_PAGE}") @Min(0L) page: Int,
                @RequestParam(defaultValue = "${PagingDefaults.THREAD_DEFAULT_SIZE}")
                @Min(1L)
                @Max(PagingDefaults.THREAD_MAX_SIZE)
                size: Int,
                authentication: Authentication?,
        ): PageResponse<ThreadSummary> {
                val ownerUsername = authSupport.requireUsername(authentication)
                return threadService.searchHidden(
                        ownerUsername,
                        query,
                        categoryIds,
                        PageRequest.of(page, size)
                )
        }

        @PostMapping
        fun createThread(
                @Valid @RequestBody request: ThreadCreateRequest,
                authentication: Authentication?,
        ): ThreadSummary {
                val ownerUsername = authSupport.requireUsername(authentication)
                return threadService.createThread(ownerUsername, request)
        }

        @GetMapping("/todo/oldest")
        fun getOldestTodoThread(
                authentication: Authentication?,
        ): ResponseEntity<ThreadDetail> {
                val ownerUsername = authSupport.requireUsername(authentication)
                val thread = threadService.getOldestThreadByCategory(ownerUsername, "todo")
                        ?: return ResponseEntity.noContent().build()
                return ResponseEntity.ok(thread)
        }

        @GetMapping("/{id}")
        fun getThread(
                @PathVariable id: String,
                @RequestParam(required = false, defaultValue = "false") includeHidden: Boolean,
                authentication: Authentication?,
        ): ThreadDetail {
                val ownerUsername = authSupport.requireUsername(authentication)
                return threadService.getThread(ownerUsername, id, includeHidden)
        }

        @GetMapping("/{id}/entries")
        fun getThreadEntries(
                @PathVariable id: String,
                authentication: Authentication?,
        ): List<EntryDetail> {
                val ownerUsername = authSupport.requireUsername(authentication)
                return threadService.getThreadEntries(ownerUsername, id)
        }

        @PatchMapping("/{id}")
        fun updateThread(
                @PathVariable id: String,
                @Valid @RequestBody request: ThreadUpdateRequest,
                authentication: Authentication?,
        ): ThreadSummary {
                val ownerUsername = authSupport.requireUsername(authentication)
                return threadService.updateThread(ownerUsername, id, request)
        }

        @DeleteMapping("/{id}")
        fun hideThread(
                @PathVariable id: String,
                authentication: Authentication?,
        ): Map<String, String> {
                val ownerUsername = authSupport.requireUsername(authentication)
                threadService.hideThread(ownerUsername, id)
                return mapOf("status" to "ok")
        }

        @PostMapping("/{id}/restore")
        fun restoreThread(
                @PathVariable id: String,
                authentication: Authentication?,
        ): ThreadSummary {
                val ownerUsername = authSupport.requireUsername(authentication)
                return threadService.restoreThread(ownerUsername, id)
        }

        @DeleteMapping("/{id}/purge")
        fun purgeThread(
                @PathVariable id: String,
                authentication: Authentication?,
        ): Map<String, String> {
                val ownerUsername = authSupport.requireUsername(authentication)
                threadService.purgeThread(ownerUsername, id)
                return mapOf("status" to "ok")
        }

        @PostMapping("/{id}/pin")
        fun pinThread(
                @PathVariable id: String,
                authentication: Authentication?,
        ): ThreadSummary {
                val ownerUsername = authSupport.requireUsername(authentication)
                return threadService.pinThread(ownerUsername, id)
        }

        @PostMapping("/{id}/unpin")
        fun unpinThread(
                @PathVariable id: String,
                authentication: Authentication?,
        ): ThreadSummary {
                val ownerUsername = authSupport.requireUsername(authentication)
                return threadService.unpinThread(ownerUsername, id)
        }

        @PostMapping("/{id}/entries")
        fun addEntry(
                @PathVariable id: String,
                @Valid @RequestBody request: EntryRequest,
                authentication: Authentication?,
        ): EntryDetail {
                val ownerUsername = authSupport.requireUsername(authentication)
                return threadService.addEntry(ownerUsername, id, request)
        }
}