# contracts/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "thredge API",
    "description": "thredge backend API (Kotlin/Spring).",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:28080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/threads": {
      "get": {
        "tags": [
          "thread-controller"
        ],
        "operationId": "list",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20,
              "maximum": 100,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseThreadSummary"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "thread-controller"
        ],
        "operationId": "createThread",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThreadCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadSummary"
                }
              }
            }
          }
        }
      }
    },
    "/api/threads/{id}/unpin": {
      "post": {
        "tags": [
          "thread-controller"
        ],
        "operationId": "unpinThread",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadSummary"
                }
              }
            }
          }
        }
      }
    },
    "/api/threads/{id}/restore": {
      "post": {
        "tags": [
          "thread-controller"
        ],
        "operationId": "restoreThread",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadSummary"
                }
              }
            }
          }
        }
      }
    },
    "/api/threads/{id}/pin": {
      "post": {
        "tags": [
          "thread-controller"
        ],
        "operationId": "pinThread",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadSummary"
                }
              }
            }
          }
        }
      }
    },
    "/api/threads/{id}/entries": {
      "get": {
        "tags": [
          "thread-controller"
        ],
        "operationId": "getThreadEntries",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntryDetail"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "thread-controller"
        ],
        "operationId": "addEntry",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntryDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/threads/{id}": {
      "get": {
        "tags": [
          "thread-controller"
        ],
        "operationId": "getThread",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeHidden",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadDetail"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "thread-controller"
        ],
        "operationId": "hideThread",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "thread-controller"
        ],
        "operationId": "updateThread",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThreadUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadSummary"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/threads/search": {
      "get": {
        "tags": [
          "thread-controller"
        ],
        "operationId": "searchThreads",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "categoryIds",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20,
              "maximum": 100,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseThreadFeedItem"
                }
              }
            }
          }
        }
      }
    },
    "/api/threads/hidden": {
      "get": {
        "tags": [
          "thread-controller"
        ],
        "operationId": "listHidden",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20,
              "maximum": 100,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseThreadSummary"
                }
              }
            }
          }
        }
      }
    },
    "/api/threads/hidden/search": {
      "get": {
        "tags": [
          "thread-controller"
        ],
        "operationId": "searchHiddenThreads",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "categoryIds",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20,
              "maximum": 100,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseThreadSummary"
                }
              }
            }
          }
        }
      }
    },
    "/api/threads/feed": {
      "get": {
        "tags": [
          "thread-controller"
        ],
        "operationId": "feed",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20,
              "maximum": 100,
              "minimum": 1
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "categoryIds",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseThreadFeedItem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/entries/hidden": {
      "get": {
        "tags": [
          "entry-controller"
        ],
        "operationId": "listHidden_1",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50,
              "maximum": 200,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseEntryDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ThreadCreateRequest": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "minLength": 1
          },
          "categoryNames": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "body"
        ]
      },
      "CategorySummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "ThreadSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "lastActivityAt": {
            "type": "string",
            "format": "date-time"
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategorySummary"
            }
          },
          "pinned": {
            "type": "boolean"
          },
          "isMarkdown": {
            "type": "boolean"
          }
        }
      },
      "EntryRequest": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "minLength": 1
          },
          "parentEntryId": {
            "type": "string"
          }
        },
        "required": [
          "body"
        ]
      },
      "EntryDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "isMarkdown": {
            "type": "boolean"
          },
          "parentEntryId": {
            "type": "string"
          },
          "orderIndex": {
            "type": "integer",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "threadId": {
            "type": "string"
          }
        }
      },
...
      "ThreadUpdateRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "categoryNames": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "PageResponseThreadSummary": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ThreadSummary"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "hasNext": {
            "type": "boolean"
          }
        }
      },
      "ThreadDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastActivityAt": {
            "type": "string",
            "format": "date-time"
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategorySummary"
            }
          },
          "pinned": {
            "type": "boolean"
          },
          "isMarkdown": {
            "type": "boolean"
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntryDetail"
            }
          }
        }
      },
      "PageResponseThreadFeedItem": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ThreadFeedItem"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "hasNext": {
            "type": "boolean"
          }
        }
      },
      "ThreadFeedItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastActivityAt": {
            "type": "string",
            "format": "date-time"
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategorySummary"
            }
          },
          "pinned": {
            "type": "boolean"
          },
          "entryCount": {
            "type": "integer",
            "format": "int32"
          },
          "isMarkdown": {
            "type": "boolean"
          }
        }
      },
      "PageResponseEntryDetail": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntryDetail"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "hasNext": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}